LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-mediaelementaudiosourcenode-interface/mediaElementAudioSourceToScriptProcessorTest-expected.txt


PASS Element Source tests completed 
FAIL Channel 0 processed some data assert_greater_than: processed data array (C0) length greater than 0 expected a number greater than 0 but got 0
FAIL All data processed correctly assert_array_approx_equals: comparing expected and rendered buffers (channel 0) lengths differ, expected 44098 got 0
